requirement.txt
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
scikit-learn==1.3.2
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.8.0
seaborn==0.13.0
joblib==1.3.2
pydub==0.25.1